refactor(proxy-core): share Xcode window message parsing
